requirements.txt
# Core
Django>=5.0,<5.1
psycopg2-binary>=2.9
redis>=5.0
django-environ>=0.11

# Image processing
Pillow>=10.0
django-imagekit>=5.0

# Import/Export & Bulk editing
django-import-export>=4.0
tablib[xlsx]>=3.5

# Admin enhancements
django-admin-sortable2>=2.1

# Async tasks
celery>=5.3
django-celery-beat>=2.5

# API
djangorestframework>=3.14
django-filter>=23.5
drf-spectacular>=0.27

# History tracking
django-simple-history>=3.4

# Dev tools
django-debug-toolbar>=4.2

# CORS for API
django-cors-headers>=4.3
